Misleading instructions for NUC
To be clearly stated the NUC image is for the hassio system disk. Not an image to create isntallation disk.
I spent 6 hours trying different computers, different SD cards and USB sticks. Googling for the solution found out was not the only one being confused here.
Just different approach how to instal when comming from the PC into Raspberry world. :-(